refactor: split tracedecay.rs edits, queries, and facts into modules
Moves the edit, query, diagnostics, and fact-store code out of
tracedecay.rs into dedicated submodules.
